+++ mobile/android/android-components/components/service/pocket/src/test/java/mozilla/components/service/pocket/ext/MappersKtTest.kt
@@ -79,7 +79,7 @@ class MappersKtTest {
 
         val result = apiStory.toLocalSpoc()
 
-        assertEquals(apiStory.flightId, result.id)
+        assertEquals(apiStory.id, result.id)
         assertSame(apiStory.title, result.title)
         assertSame(apiStory.url, result.url)
         assertSame(apiStory.imageSrc, result.imageUrl)